+++ src/main/webapp/site/src/app/contact/contact-form/contact-form.component.scss
@@ -29,3 +29,8 @@
 .contact__form {
   width: 800px;
 }
+
+.recaptcha-container {
+  display: flex;
+  justify-content: center;
+}
