+++ src/main/webapp/site/src/app/login/login-home/login-home.component.scss
@@ -1,1 +1,7 @@
 
+
+.recaptcha-container {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+}
